refactor(background): extract the durable Organize job host
`src/background/index.ts` was still 3422 lines, and half of it was one
subsystem: the durable Organize job host. Moving it to
`src/background/organize-job-host.ts` brings the worker's composition root to
1548 lines.

The moved code is the connection registry, trace coordinator, analysis
scheduler, apply pump and its recovery, the `bgsm-agent-organize-job` port
protocol, and the publication and control helpers that address a reconnected
page. These form one late-bound cycle — the controller schedules through the
scheduler while the scheduler drives the controller — which is why the previous
split left them alone. Constructing them together inside
`createOrganizeJobHost` makes that cycle local to the module instead of
requiring a forward-declared `let` at worker top level.

The boundary is narrow in both directions and was measured, not assumed. The
host reaches back through four `deps` entries (`jobQueue`,
`broadcastDataChanged`, `createRuntimeProvider`, `observeExecutionEvent`); the
worker reaches in through the two entries the factory returns (`install`,
`publishAgentSessionDeleted`). All 1701 moved statements are verbatim apart from
the two dependency callsites that now go through `deps`.

`background-watch-integration-contract` anchored its storage-listener contract
on `\n});\nconst organizeJobRunConnections`, which the move invalidated. It is
re-anchored on the listener's own `areaName !== 'local'` guard so that moving
neighbouring code cannot silently drop the assertion.

Constraint: the Organize controller and scheduler are mutually dependent, so they must be constructed in one scope; and the shared serialized queue must stay shared, because an Apply chunk racing a Stars sync would corrupt tag writes.
Rejected: passing the whole worker module as a context object, which would recreate the coupling the extraction removes; and splitting the host further into port/publication/wiring modules, since the port listener alone needs 8 pieces of wiring state and 16 publication helpers, so the seam would be wider than the file.
Confidence: high
Scope-risk: broad
Directive: 只抽 Organize host。
Tested: pnpm typecheck; pnpm test:logic (2624); pnpm test:integration (21); pnpm test:regressions (741); pnpm test:runtime; the six source-text contract suites (56 tests); pnpm build; pnpm test:smoke passed all 12 full-product Chrome scenarios with zero page, background, or uncaught errors. A statement-level diff confirmed every moved line is verbatim and the retained part lost only now-unused imports.
Not-tested: pnpm test:runtime:organize-job-host, which fails identically on baseline HEAD — same primaryCode trusted_origin_deletion_dismiss_evidence_failed and the same 11 unexpected POST requests — so it is a pre-existing failure this change neither causes nor fixes.
